move append_path_params
